Images working in the new ui

+++ src/mouse.rs
@@ -24,34 +24,35 @@ fn mouse_position(
 ) {
     // get the camera info and transform
     // assuming there is exactly one main camera entity, so query::single() is OK
-    let (camera, camera_transform) = q_camera.single();
+    if let Ok((camera, camera_transform)) = q_camera.get_single() {
+        // assuming camera has render target window (instead of Image)
+        let window_id = match camera.target {
+            RenderTarget::Window(win_id) => win_id,
+            _ => panic!("expecting camera to have rendertarget window"),
+        };
 
-    // assuming camera has render target window (instead of Image)
-    let window_id = match camera.target {
-        RenderTarget::Window(win_id) => win_id,
-        _ => panic!("expecting camera to have rendertarget window"),
-    };
+        // get the window that the camera is displaying to
+        let wnd = wnds.get(window_id).unwrap();
 
-    // get the window that the camera is displaying to
-    let wnd = wnds.get(window_id).unwrap();
+        // check if the cursor is inside the window and get its position
+        if let Some(screen_pos) = wnd.cursor_position() {
+            // get the size of the window
+            let window_size = Vec2::new(wnd.width() as f32, wnd.height() as f32);
 
-    // check if the cursor is inside the window and get its position
-    if let Some(screen_pos) = wnd.cursor_position() {
-        // get the size of the window
-        let window_size = Vec2::new(wnd.width() as f32, wnd.height() as f32);
+            // convert screen position [0..resolution] to ndc [-1..1] (gpu coordinates)
+            let ndc = (screen_pos / window_size) * 2.0 - Vec2::ONE;
 
-        // convert screen position [0..resolution] to ndc [-1..1] (gpu coordinates)
-        let ndc = (screen_pos / window_size) * 2.0 - Vec2::ONE;
+            // matrix for undoing the projection and camera transform
+            let ndc_to_world =
+                camera_transform.compute_matrix() * camera.projection_matrix.inverse();
 
-        // matrix for undoing the projection and camera transform
-        let ndc_to_world = camera_transform.compute_matrix() * camera.projection_matrix.inverse();
+            // use it to convert ndc to world-space coordinates
+            let world_pos = ndc_to_world.project_point3(ndc.extend(-1.0));
 
-        // use it to convert ndc to world-space coordinates
-        let world_pos = ndc_to_world.project_point3(ndc.extend(-1.0));
+            // reduce it to a 2D value
+            let world_pos: Vec2 = world_pos.truncate();
 
-        // reduce it to a 2D value
-        let world_pos: Vec2 = world_pos.truncate();
-
-        mouse_position.0 = world_pos;
+            mouse_position.0 = world_pos;
+        }
     }
 }
